Created Card, Deck, Hand, Chips Classes
Created four classes; Card, Deck, Hand, and Chip.  Card class gives each card a suit, rank, and value.  The Deck class creates of deck of 52 cards and adds them to a list.  The Deck class also shuffles deck and deals one card using the .pop function.  The Hand class creates a list of cards in the player's hand, figures out the value of the cards in the player's hand, and adjusts for Aces (whether they're worth 11 or 1).  The Chips class creates the total monies the player has to use in the game and subtracts or adds to the players total depending on the bet.